sqlite3* lemon: updated to 3.45.1
version 3.45.1 (2024-01-30):

Restore the JSON BLOB input bug, and promise to support the anomaly in subsequent releases, for backward compatibility.
Fix the PRAGMA integrity_check command so that it works on read-only databases that contain FTS3 and FTS5 tables. This resolves an issue introduced in version 3.44.0 but was undiscovered until after the 3.45.0 release.
Fix issues associated with processing corrupt JSONB inputs:
Prevent exponential runtime when converting a corrupt JSONB into text.
Fix a possible read of one byte past the end of the JSONB blob when converting a corrupt JSONB into text.
Enhanced testing using jfuzz to prevent any future JSONB problems such as the above.
Fix a long-standing bug in which a read of a few bytes past the end of a memory-mapped segment might occur when accessing a craftily corrupted database using memory-mapped database.
Fix a long-standing bug in which a NULL pointer dereference might occur in the bytecode engine due to incorrect bytecode being generated for a class of SQL statements that are deliberately designed to stress the query planner but which are otherwise pointless.