go-module, go-package: set GOTOOLCHAIN=local
From version 1.21 on, Go has built-in support for additional toolchains.
If the installed Go version is too old, it can download and use a newer
version on the fly during the build.

See https://tip.golang.org/doc/toolchain for details.

Needless to say, we don't want this to happen during a pkgsrc build.
This setting disables the behavior and always uses the bundled toolchain.