py-gevent: updated to 23.9.1
23.9.1 (2023-09-12)
Bugfixes
Require greenlet 3.0 on Python 3.11 and Python 3.12; greenlet 3.0 is recommended for all platforms. This fixes a number of obscure crashes on all versions of Python, as well as fixing a fairly common problem on Python 3.11+ that could manifest as either a crash or as a SystemError.

23.9.0.post1 (2023-09-02)
Fix Windows wheel builds.
Fix macOS wheel builds.

23.9.0 (2023-09-01)
Bugfixes
Make gevent.select.select accept arbitrary iterables, not just sequences. That is, you can now pass in a generator of file descriptors instead of a realized list. Internally, arbitrary iterables are copied into lists. This better matches what the standard library does. Thanks to David Salvisberg.
On Python 3.11 and newer, opt out of Cython’s fast exception manipulation, which may be causing problems in certain circumstances when combined with greenlets.
On all versions of Python, adjust some error handling in the default C-based loop. This fixes several assertion failures on debug versions of CPython. Hopefully it has a positive impact under real conditions.

Make gevent.pywsgi comply more closely with the HTTP specification for chunked transfer encoding. In particular, we are much stricter about trailers, and trailers that are invalid (too long or featuring disallowed characters) forcibly close the connection to the client after the results have been sent.