tex-albatross{,-doc}: update to 0.5.1
This release does not contain the JNA code for Apple's darwin platform anymore
due to code signing issues. While we could not test the effectiveness and
usability of the result, the MacTeX maintainer already uses this strategy to
distribute albatross. If there are any issues, please report and we will revert
until we receive contributions with a proper fix.