textproc/ruby-asciidoctor: update to 2.0.21
2.0.21 (2024-02-20)

This is a patch release in the 2.0.x release line, primarily to address
compatibility with Ruby 3.3.  While development of Asciidoctor 2.1 is
beginning to pick up again, we extracted several other important
improvements and bug fixes from main to include in this patch release.

The most critical fix in this release is to make Asciidoctor's logger
compatible with the logger API in Ruby 3.3.  The code is now tested against
Ruby 3.3 on each change.  Nearly as important, this release also turns off
Ruby's system-dependent newline conversion when writing files so that
Asciidoctor's output is consistent between *nix and Windows.

Numerous changes were made to correct AsciiDoc parsing.  A nested dlist
attached using a list continuation no longer consumes a detached paragraph.
And an attached block is correctly attached to a nested dlist when that
dlist is offset from its parent by an empty line.  The secondary and
tertiary terms on the indexterm macro are preserved when the primary term is
quoted and contains an equals sign.  And a bare URL enclosed in angle
brackets acts as unconstrained syntax that stops at the first occurrence of
a closing angle bracket.  When includes are not enabled, any spaces in the
include target are preserved when generating the fallback link.

Several changes were made to the DocBook output to address compliance.  The
abstract is moved inside the <info> tag (applies to both the abstract in the
preamble as well an an abstract for a book part).  The scaledwidth and scale
attributes are now supported on an inline macro, like with a block image
macro.

A few changes to the manpage output were made.  Cells in the head row are
styled as bold.  Paragraph breaks in a normal table cell are preserved.

Finally, a few refinements to the default stylesheet were applied.  The
extra border below the doctitle when the sidebar toc is collapsed into main
content area was removed.  And a text decoration was added to the active
footnote number link in the footnotes list.

Compliance

* Turn off system-dependent newline conversion when writing files; don't
  convert line feeds to system-dependent newline (#4550)

* Support logger in Ruby 3.3 by instantiating super class (#4493) (@mtasaka)

* Add support for scaledwidth and scale attributes on inline image macro in
  DocBook output (#4552)

* Update latest Ruby to 3.3 in CI workflows

Improvements

* Change title to doctitle in warning message about use of abstract to make
  subject more clear

* Modify default stylesheet to add text decoration to active footnote number
  link in footnotes list (#4530) (@Larhzu)

Bug Fixes

* Nested dlist attached using list continuation should not consume detached
  paragraph (#3535)

* Don't break nested dlist with attached block if offset from parent list by empty line (#3693)

* Preserve paragraph breaks in normal table cell in manpage output (#4481)

* Style cells in head row as bold in manpage output (#4490)

* Escape spaces in include target (using inline passthrough) when generating
  link from include directive (#4461)

* Move abstract inside info tag in DocBook output (#3602)

* Honor secondary and tertiary terms on indexterm macro when primary term is
  quoted and contains an equals sign (#3652)

* Remove extra border below doctitle when sidebar toc is collapsed into main
  content area (#4523)

* Treat bare URL enclosed in angle brackets as unconstrained syntax; only
  match until closing angled bracket (#4468)

* Allow URL enclosed in angled brackets syntax to be escaped using backslash
  (#4468)

Credits

A very special thanks to all the awesome supporters of the Asciidoctor
OpenCollective campaign, who provide critical funding for the ongoing
development of this project.