libfontenc: update to 1.1.8.
This release includes build system cleanups and adds "e" to the fopen()
flags to avoid leaking fd's when calling applications exec another program.

Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow it to
work with files whose timestamps are later than January 19, 2038 on
32-bit platforms, but this has not been tested.   This only applies to
non-compressed files opened by this library - compressed files are
handled by libz and depend on whether it was built with post-Y2038 support.

Alan Coopersmith (5):
      configure: raise minimum autoconf requirement to 2.70
      Remove "All rights reserved" from Oracle copyright notices
      configure: Use LT_INIT from libtool 2 instead of deprecated AC_PROG_LIBTOOL
      Set close-on-exec when opening files
      libfontenc 1.1.8